shamu: Remove liboemcrypto.so
Removing liboemcrypto.so forces Widevine L1 DRM to fallback to L3

Ethan Chen:
I've got some confirmation L1 playback will not  work on devices
with the bootloader state set to "unlocked", due to a check that
happens inside the widevine trustzone app.
So we're not really able to ship liboemcrypto regardless
of whatever is going on here.

Change-Id: I8a2b773c0b65dbfc1acd348333da2391212cc45d